Robert's Edit September 9 2024
Finished the Resident Module except for the Edit camera capture.

Fixed the footer mixing with the body

Make the table dynamic and update the table with every operation.

Make the deleted records visible in the front end.

Can now capture images from the camera.

Makes all the PHP script for Add, Edit ,Delete, Pagination, in one file.

Now has search for records and automatically search for records upon type of more than 2 characters including the deleted records.

Put a switch to see the deleted records.

 Put all the Operation Javascripts in one file.

DIsables the File upload input when the camera was used.

Deletes the residents.php table body to be filled by the Javascript and AJAX.

Fixes the </div> in the HTML file attached.

 Put a copy of the current database as of this writing.

The residents table now has pagination even in the deleted records.

Put the resident table body and pagination in a separate file to be fetched by AJAX.

It makes an error-catching for the Image Upload.

Deletes unnecessary files and redundant codes deleted.

Makes the placeholder image offline.

Makes a div in the modal for the Camera feed and image preview.

Edit the Database for more applicable design for the barangay.

Add more comments for better understanding.

Put an if statement for the displayimagedata.js to differentiate View and Edit modal.

Adds functions to the JavaScript for frequent operations.