Roberts Edit Sepetember 15 2024
Change the Add Resident to New Resident.

Creates a sidebar active highlight.

Finishes the View Resident in the Documents module

Put the expiration date and the model of the documents in the table.

Fixes the template for the status symbol

Replaces the Resident Is a voter whith check and x marks

Converts the Javascript of the template to Jquery

Added the current database