Robert Edit November 14 2024
Blotters almost complete just some polishing.

Put security mesures in accessing operation files in includes.

Added blotter schedule manager

Changes the Add blotter to New Blotter

Edited the database for additional information and functions.

Fixes errors in the Non Resident Module and the edit of the documents module.

Created a Login page for the login.

Started the User management module

Minor edits to the css for the login page.

Removes redundant codes in the operation files and template.