Robert Edit November 18 2024
Fixes Major problems with Resident and Non Resident (Except the file upload of camera of Non resident).

Removes the UTF8 Decode from the certificates.

Replaces the strtoupper to mb_strtoupper.

Moves deleted residets sql query to view

Fixes the full calendar details display

Removes the user in the sidebar.

Replaces Select  something in the select to just select

Fixes the camera feed of the non resident

Puts a header utf8 text/html to every generate file of the certificate.

Combine the pagination of the 2 table in the modal of resident and non resident

Removes redundant codes

Continuation of the user management