Robert Edit November 23 2024
Fixes the Blotter hit badge in the Resident and Non Resident View

Fixes the code of the view image in the resident and non resident module

Make the login username case sensitive

It now has auto logout function after 5 mins of inactivity

Now has audit trail to check ever add, edit and delete of the user.

Increase the max resolution of the image upload.

Remove unnesesary files and code