Robert Edit November 27 2024
Removes redundancy to the code

Fixes the certificate generation errors and adds more error catching for easier trobleshooting.

Started the audit trail.

Can now generate reports for the documents.

Make the add resident and non resident in the create blotters.

Fixes the blotters won't update the entry in the database despite replying success.

Added additional statistics to the homepage.

Make the check hit works in the profile of resident and non resident.

Remove unnecessary elements in the template

Move the expiration to the tbl_docu_request.

Created a settings page for the brgy details and images.

Created a page for user management.