Robert Edit December 10 2024
Completed the Enforcement of the account restrictions.

Added the Add Resident and Non Resident form in the Create blotter page.

Completed the audit trail

Removes the redundant codes

Makes the view modal of the blotter rather than simple sweet alerts in the blotter schedule page.

Enforce the login on certain pages.

Consolidate codes that can be used in else where.